Separate the policy step into 2: name/description and roles selection
